# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Archiver API",
    "version": "1.0.0",
    "description": "REST API for Open Archiver — an open-source email archiving platform. All authenticated endpoints require a Bearer JWT token obtained from `POST /v1/auth/login`, or an API key passed as a Bearer token.",
    "license": {
      "name": "SEE LICENSE IN LICENSE",
      "url": "https://github.com/LogicLabs-OU/OpenArchiver/blob/main/LICENSE"
    },
    "contact": {
      "name": "Open Archiver",
      "url": "https://openarchiver.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from `POST /v1/auth/login`. Pass as `Authorization: Bearer <token>`."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API key generated via `POST /v1/api-keys`. Pass as `X-API-KEY: <key>`."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication is required or the token is invalid/expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated user is not allowed to perform this action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Forbidden"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "An unexpected error occurred on the server.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Internal server error"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error description.",
            "example": "An error occurred."
          }
        },
        "required": [
          "message"
        ]
      },
...
      "SearchResults": {
        "type": "object",
        "properties": {
          "hits": {
            "type": "array",
            "description": "Array of matching archived email objects, potentially with highlighted fields.",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          },
          "processingTimeMs": {
            "type": "integer",
            "description": "Meilisearch query processing time in milliseconds.",
            "example": 12
          }
        },
        "required": [
          "hits",
          "total",
          "page",
          "limit",
          "totalPages",
          "processingTimeMs"
        ]
      },
...
      "QueueCounts": {
        "type": "object",
        "properties": {
          "active": {
            "type": "integer",
            "example": 0
          },
          "completed": {
            "type": "integer",
            "example": 56
          },
          "failed": {
            "type": "integer",
            "example": 4
          },
          "delayed": {
            "type": "integer",
            "example": 0
          },
          "waiting": {
            "type": "integer",
            "example": 0
          },
          "paused": {
            "type": "integer",
            "example": 0
          }
        },
        "required": [
          "active",
          "completed",
          "failed",
          "delayed",
          "waiting",
          "paused"
        ]
      },
      "QueueOverview": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ingestion"
          },
          "counts": {
            "$ref": "#/components/schemas/QueueCounts"
          }
        },
        "required": [
          "name",
          "counts"
        ]
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true,
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "initial-import"
          },
          "data": {
            "type": "object",
            "description": "Job payload data.",
            "example": {
              "ingestionSourceId": "clx1y2z3a0000b4d2"
            }
          },
          "state": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "failed",
              "delayed",
              "waiting",
              "paused"
            ],
            "example": "failed"
          },
          "failedReason": {
            "type": "string",
            "nullable": true,
            "example": "Error: Connection timed out"
          },
          "timestamp": {
            "type": "integer",
            "example": 1678886400000
          },
          "processedOn": {
            "type": "integer",
            "nullable": true,
            "example": 1678886401000
          },
          "finishedOn": {
            "type": "integer",
            "nullable": true,
            "example": 1678886402000
          },
          "attemptsMade": {
            "type": "integer",
            "example": 5
          },
          "stacktrace": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "returnValue": {
            "nullable": true
          },
          "ingestionSourceId": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "description": "Shorthand copy of `failedReason` for easier access.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "data",
          "state",
          "timestamp",
          "attemptsMade",
          "stacktrace"
        ]
      },
      "QueueDetails": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ingestion"
          },
          "counts": {
            "$ref": "#/components/schemas/QueueCounts"
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "currentPage": {
                "type": "integer",
                "example": 1
              },
              "totalPages": {
                "type": "integer",
                "example": 3
              },
              "totalJobs": {
                "type": "integer",
                "example": 25
              },
              "limit": {
                "type": "integer",
                "example": 10
              }
            },
            "required": [
              "currentPage",
              "totalPages",
              "totalJobs",
              "limit"
            ]
          }
        },
        "required": [
          "name",
          "counts",
          "jobs",
          "pagination"
        ]
      },
...
    }
  },
  "paths": {
...
    "/v1/search": {
      "get": {
        "summary": "Search archived emails",
        "description": "Performs a full-text search across indexed archived emails using Meilisearch. Requires authentication.",
        "operationId": "searchEmails",
        "tags": [
          "Search"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "keywords",
            "in": "query",
            "required": true,
            "description": "The search query string.",
            "schema": {
              "type": "string",
              "example": "invoice Q4"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number for pagination.",
            "schema": {
              "type": "integer",
              "default": 1,
              "example": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of results per page.",
            "schema": {
              "type": "integer",
              "default": 10,
              "example": 10
            }
          },
          {
            "name": "matchingStrategy",
            "in": "query",
            "required": false,
            "description": "Meilisearch matching strategy. `last` returns results containing at least one keyword; `all` requires all keywords; `frequency` sorts by keyword frequency.",
            "schema": {
              "type": "string",
              "enum": [
                "last",
                "all",
                "frequency"
              ],
              "default": "last"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResults"
                }
              }
            }
          },
          "400": {
            "description": "Keywords parameter is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/jobs/queues": {
      "get": {
        "summary": "List all queues",
        "description": "Returns all BullMQ job queues and their current job counts broken down by status. Requires authentication.",
        "operationId": "getQueues",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of queue overviews.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "queues": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/QueueOverview"
                      }
                    }
                  },
                  "example": {
                    "queues": [
                      {
                        "name": "ingestion",
                        "counts": {
                          "active": 0,
                          "completed": 56,
                          "failed": 4,
                          "delayed": 3,
                          "waiting": 0,
                          "paused": 0
                        }
                      },
                      {
                        "name": "indexing",
                        "counts": {
                          "active": 0,
                          "completed": 0,
                          "failed": 0,
                          "delayed": 0,
                          "waiting": 0,
                          "paused": 0
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/jobs/queues/{queueName}": {
      "get": {
        "summary": "Get jobs in a queue",
        "description": "Returns a paginated list of jobs within a specific queue, filtered by status. Requires authentication.",
        "operationId": "getQueueJobs",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "queueName",
            "in": "path",
            "required": true,
            "description": "The name of the queue (e.g. `ingestion` or `indexing`).",
            "schema": {
              "type": "string",
              "example": "ingestion"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter jobs by status.",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "completed",
                "failed",
                "delayed",
                "waiting",
                "paused"
              ],
              "default": "failed"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed view of the queue including paginated jobs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueDetails"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Queue not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
...
}